WaRP: Removed I2C2 configuration from bootloader
WaRP would hang on boot if touchscreen was not connected
due to perceived contention on the I2C2 bus when a
touchscreen is not connected. Because touch and I2C2 is
not used in the bootloader, this configuration has been
removed rather than enabling internal iMX6 pull-ups
that could potentially conflict with touchscreen pull-ups
in the future.

Signed-off-by: Jacob Postman <jacob@revolution-robotics.com>